libpressio version 0.88.0
Major Changes

+ **Breaking Change** pressio_thread_safety and pressio_dtype can now be
  stored in pressio_option structures.  This change improves ergonomics
  and typesafety for options using these types.  Unfortunately, this
  breaks the API for users of pressio:thread_safe.
+ normalize_dims now allows the user to specify the size and fill for
  dimension normalization.  Once fully supported, all compressors can
  trivially support higher dimensions by concatenation.

Minor Changes

+ Python bindings now load 3rd party compressors by default

Bug Fixes

+ many_independent_threaded previously didn't check when updating its
  array of metrics results if the previous result were empty.  Now it
  does so meaningful results are returned.
+ Missing header in data.h was added

+++ src/plugins/compressors/log_transform.cc
@@ -60,7 +60,7 @@ y[0] = log(x[0]);
   pressio_options get_configuration_impl() const override {
     pressio_options opts;
     opts.copy_from(meta->get_configuration());
-    set(opts, "pressio:thread_safe", static_cast<int32_t>(get_threadsafe(*meta)));
+    set(opts, "pressio:thread_safe", get_threadsafe(*meta));
     set(opts, "pressio:stability", "experimental");
     return opts;
   }
