+++ src/plugins/compressors/sz_auto_plugin.cc
@@ -30,7 +30,7 @@ class sz_auto_plugin: public libpressio_compressor_plugin {
 
     struct pressio_options get_configuration_impl() const override {
       struct pressio_options options;
-      set(options, "pressio:thread_safe", static_cast<int>(pressio_thread_safety_multiple));
+      set(options, "pressio:thread_safe", pressio_thread_safety_multiple);
       set(options,"pressio:stability", "experimental");
       return options;
     }
